- improved functionality of "Element Get Settings" and "Element Apply Settings" - Added functionality for Duplicate Element - Duplicate now creates a new element instead of referencing the original - Added icons for chain elements - Added enum for chain elements - tweaked ui - Added functionality to edit the save path via ui

+++ MovementScriptGenerator/Modules/Moves/Spiral.cs
@@ -10,7 +10,7 @@ namespace MovementScriptGenerator
         public float EndDistance { get; }
         public float HorizontalRot { get; }
         public float VerticalRot { get; }
-        public float SpiralAmmount { get; }
+        public int SpiralAmmount { get; }
         public bool SpiralClockwise { get; }
         public float StartHold { get; }
         public float EndHold { get; }
